Add CBB_add_asn1_[u]int64_with_tag.
CBB_add_asn1_uint64 doesn't work if you're encoding an implicitly-tagged
INTEGER. Take a leaf from Go cryptobyte and add a "with tag" variant,
rather than a "contents" variant, which is a little more convenient to
use. It also avoids us having to decide how to name the contents field.

Change-Id: I6072e55017230c513577c44c5a7ed86e778255b3