Enable -Wstring-concatenation and silence warning.
Newer versions of Clang have a warning to detect "suspicious" uses of
string concatenation, where they thing a comma or so was missing. It
flags a false positive in x509_test.cc, which we can silent with
parentheses. Fuchsia builds with this warning enabled, so enable it to
catch future instances.

I couldn't find official documentation on when this was added, but
empirically it's in my clang-12 but not my clang-11. That's recent
enough that adding a version check seems prudent. Unfortunately,
version-detecting Clang is complex because AppleClang uses completely
different versions. There's a handy table on Wikipedia that maps them.

Change-Id: I503c21d39bb5c68dda9bda6da693c7208f3af561