Fix integer overflow in OPENSSL_gmtime_adj
OpenSSL sent integer parameters as this, and the multiplication
here ends up being done as an integer.  Since we support values
up to year 9999, it is possible for someone to pass in a number of
days to the "adj" function to adjust a base time far enough to
overflow a 32 bit integer.

Change-Id: Iedfc33d8bf90d70049f99897df1d193fb29805d0