Rename and tidy up x509v3_name_cmp.
First, rename to x509v3_conf_name_matches and flip the result value. We
don't need to preserve the positive vs negative return of strncmp here.
The rename is because "name" can mean so many things in the context of
X.509. Here, it's specifically the name of a CONF_VALUE.

Finally, fix it to be size_t-clean.

Bug: 516
Change-Id: I1c3039d9c6ce70cde669e07f943ad1e25fb49dc1