Give cipher tests a common entrypoint.
The main FileTest callback didn't support invalid ciphertexts, while the
Wycheproof one did but reimplemented a lot of logic to discover
variations.

Rework this a bit so we have a single TestCipher function that's
agnostic to the FileTest setup. The Operation enum gains a
kInvalidDecrypt mode to specify that the ciphertext is invalid. As part
of this, we tighten up the test

This is to make the later changes easier:

1. We don't have any negative tests for EVP_CIPHER's tag check at all!
   We only test it for EVP_AEAD right now. Now TestCipher can express
   this.

2. The weird EVP_Cipher API has no tests. Its calling convention is very
   surprising and will warrant some particular expectations. Previously,
   kInvalidDecrypt was implemented by checking if DoCipher failed, which
   makes it harder to embed success/failure-specific assertions. By
   moving everything to a common entry

Along the way, I've made each function responsible for running through
the input once, with separate TestCipherAPI and TestLowLevelAPI helpers.
Otherwise we have to reset the spans and keep track of the state of
intermediate values.

Bug: 494
Change-Id: Ieab87c0c76586aefeb596cc90edd4910ff1b962f