Drop the workaround for the third-party Android CMake toolchain.
When we first added Android support to the CMake build, we used a
third-party toolchain file, https://github.com/taka-no-me/android-cmake.
That project hasn't been updated since 2015 and, among other things,
doens't support Armv8.

In 2017, we replaced it with the NDK's build-in toolchain file, which no
longer needed a particular workaround for ASM flags. See
https://boringssl-review.googlesource.com/c/boringssl/+/24165 and
https://boringssl-review.googlesource.com/c/boringssl/+/24164

Drop this workaround now.

Change-Id: I4af78d641b0c9b9608aaaa5ca31b9e80de024ab2