Compute policy cache as part of x509v3_cache_extensions.
This fixes a race condition in the X.509 policy logic. Setting up the
policy cache can write to x->ex_flags, but verification can read from
it.

There's a red herring here that we, when porting away from the old
locks, put the policy cache under one global lock while
x509v3_cache_extensions (which sets up x->ex_flags) under x->lock.
However, simply moving everything to x->lock (which matches upstream)
does not actually solve the problem. It is a race condition either way.

This is because x->ex_flags is *not* guarded by x->lock. It, along with
a host of other state, is call_once-initialized by
x509v3_cache_extensions. All code then assumes that, after
x509v3_cache_extensions has run, all relevant values can be freely accessed
without initialization. Critical to this assumption is that all those
values are immutable once created.

The policy cache breaks this assumption by modifying it afterwards. I've
added a test which reproduces this, both with the global lock and the
red herring x->lock non-fix. The latter takes a few tries for TSan to
find it, but it does flag it.

Instead, this change moves policy cache initialization into
x509v3_cache_extensions. It preserves the use of separate EXFLAG_INVALID
and EXFLAG_INVALID_POLICY bits, so verifications without policy checking
will continue to not notice policy syntaxes. (Arguably there is a bug
here around critical policy constraints extensions...)

One of the consequences of this change is that the policy cache setup
code (but not tree building) is now exposed to all callers, not just the
few that enable policy verification. Hopefully the preciding changes
will provide us some test coverage, fuzzing coverage, as well as a
general review pass to be more confident in it, at least relative to the
low baseline of confidence we have in crypto/x509 itself.

Change-Id: I85ebe4c838d64679087819a1eb53cbdc801e64c4