Const-correct various policy tree functions.
As part of this, I've intentionally *removed* const on
x509_policy_cache_find_data. Although the compiler does not notice, that
function returns a non-const pointer and also mutates cache->data to
re-sort it if necessary. It should never be called outside of cache
creation (and indeed it isn't).

I've left a TODO for one place that will require larger changes to fix
in a follow-up CL.

Change-Id: Ibb47598800e7912dd3fbefc805793985e55c8a8d