Remove the tree parameter to x509_level_add_node and const-correct.
x509_level_add_node is a mess of a function. The return value is owned
or non-owning depending on whether level is NULL. It takes ownership of
data or not depending on whether tree is NULL.

This CL removes the second of these effects but leaves the first.
Removing the second allows us to const-correct the data parameter, which
is important for thread safety. Instead, make saving non-cache nodes in
the tree a separate operation for callers to perform.

The first effect is just documented. But, in doing so, I've fixed a few
bugs where it comes up:

- On malloc failure, extra nodes weren't freed.
- Nothing checked for malloc failure in x509_level_add_node.
- exnode_free is renamed to clarify what's going on here.

This code is covered by the anyPolicy tests just added, so that's some
reason to believe it's correct.

Change-Id: I3c3a72d7d7776e4b829fffc49fd4841a2a44ac5f