Avoid quadratic loop in x509_policy_cache_set_mapping.
In every loop iteration, the function potentially adds to the cache, and
also queries from the cache. We'd need a proper map, but currently just
use a sorted list.

Instead, defer the insertions, so we can query the existing cache in
O(log N) time without a re-sort. We still need to query synthesized
entries, but if we add entries in the right order, we only need to check
the most recent synthesized entry.

Policy verification itself scales far from well, but that's not enabled
by default. A subsequent change will make policy cache creation always
enabled, for the sake of fixing a race condition, so we should keep that
subquadratic.

Change-Id: Iecc66ff8bd957cd2a490b4a3736f66b077e15183