Add CBS_get_u64_decimal.
ASN1_generate_v8 has a number of calls to strtoul. strtoul has two
problems for that function.

First, strtoul keeps reading until NUL, but all the functions in that
file act on pointer/length pairs. It's fine because the underlying
string is always NUL-terminated, but this is fragile.

Second, strtoul is actually defined to parse "-1" as
(unsigned long)(-1)! Rather than deal with this, extract the decimal
string parser out of the OID parser as a CBS strotul equivalent.

Change-Id: I1b7a1867d185e34e752be09f8c8103b82e364f35