Reduce caller requirements on X509V3_CTX.
Although one needs to initialize X509V3_CTX in two parts, some callers
forget to to this. Since all the entrypoints take a CONF anyway, and
always match them, just implicitly initialize the CONF half of the
X509V3_CTX with the provided one.

Also allow X509V3_CTX to be NULL. Some code checks for it and some
doesn't. Enough doesn't that it really cannot be considered to work, but
I've seen one caller mistakenly do this. Since we have to copy the
X509V3_CTX anyway, allow it to be NULL and fill in an empty one when
omitted.

Update-Note: If using different CONFs in the X509V3_CTX and the function
parameter, the function parameter is now always used. No callers do
this, and it's somewhat arbitrary which is used. (The generic code
always uses the one in ctx. The @section syntax uses the parameter. Then
the per-extension callbacks use the ctx.)

Change-Id: I9fc15a581ea375ea06c4b082dcf0d6360be8144f