Add some tests for optional and default ASN1_BOOLEAN.
ASN1_BOOLEAN has these ASN1_FBOOLEAN and ASN1_TBOOLEAN variants that
behave slightly strangely. Add some tests to ensure we don't break them
in the rewrite.

In doing so, I noticed a bug: ASN1_BOOLEAN canonically represents TRUE
as 0xff, to match DER. But ASN1_TBOOLEAN is initialized with it->size,
which is 1, not 0xff. Fix it to be 0xff.

Bug: 548
Change-Id: I4e7fdc2a3bc87603eaf04a7668359006a1480c2e