Const-correct sk_FOO_cmp_func [DO NOT SUBMIT YET]
This is a double-pointer and both layers should be const. This matches
OpenSSL 1.1.1, so in addition to being more const-correct, we're more
OpenSSL-compatible.

Update-Note: Anything that defines a comparison function would need to
fix the type signature. I found only one external caller, Envoy, that
defines it. [TODO LINK TO PR] fixes it. (That we hadn't run into the
upstream incompatibility suggests this is just not a feature folks use
outside the library much.)

Fixed: 498
Change-Id: I8f07ff42215172aa65ad8819acf69b63d6d8e54c