Maintain a frame pointer in aesni-gcm-x86_64.pl.
Some profiling systems cannot unwind with CFI and benefit from having a
frame pointer. Since this code doesn't have enough register pressure to
actually need to use rbp as a general register, this change tweaks
things so that a frame pointer is preserved.

Performance difference is within measurement noise.

This does not create a stack frame for internal functions so
frame-pointer unwinding may miss a function or two, but the broad
attribution will be correct.

Change by Clemens Fruhwirth.

Bug: b/33072965
Change-Id: I86cdcdac2ba2d104c422e5b9a67b2d46473f44a5